refactor(kamn-node): extract bootstrap report builder module (#2490)
